Add latex support for BendingOutputs component
Fixed a bug where the calculated number would not be displayed in a
MathJax component. When a new state was set and the content of the page
changed, MathJax did not rerender. While state was changed in memory it
was not displayed on the screen. The following props had to be
passed to the MathJax element hideUntilTypeset={"first"} inline dynamic