add react-hook-form library to handle froms
change managment of input data. Before the change I was using uesRef to
get access to the dom-nodes of the input elements in order to get the
associated values. Now react-hook-forms handles the submit. After
submiting we get an object containing key value pairs of the input label
and the input value